Remove __array__ from Matrix base class
Keeping this causes the Matrix to be converted to a ndarray when broadcasted
to another array or a number. This interferes with simple scalar
multiplications: While `matrix * 1` returns a matrix, `1 * matrix` fails in
trying to return an array.

Besides, on a strict mathematical standpoint it is highly non-intuitive for
a matrix to turn into an array out of blue skies.