Fix CG(Eye()) issue
In the discussion to #26 an incompatibility issue was reported when using CG
on an pure Eye Matrix, resulting in zeroing the input data vector. This was
resolved by forcing the temporary residual array object within CG to differ
from the input array. As Eye is currently the only Matrix directly returning
the input array object from its transforms the scope of the problem is quite
limited. However, this could also be an issue for Partial matrices, too,
which return views into potential other input references.

Just stay safe when writing into transform output arrays in the future!