Introduce array squeeze with dimension checks
core.cmath._arrSqueeze()
core.cmath._arrSqueezedCopy()

These routines make sure a given vector or number is converted to a ndarray
while removing empty axes and ensuring there is at least one dimension.
The second also ensures the output array points to a unique copy of the data
pointed to by the input array.

Use these routines to make sure that an array fulfils minimum requirements
and ease some debugging hassle. The use of these routines over a simple copy()
during initialization of a class' instance must be enforced as this is the
place where a central behaviour flag can be put in the future to weigh up
between making instances as robust as possible and to (potentially) save lots
of loads of memory.